Display all PDBs released a week from today.
Previously only new PDBs that ended up in one of the NR classes were returned. This ignored PDB files without complete nucleotides, which was confusing.
Since the homepage is cached and the cache is updated weekly, the homepage should capture the state of the system at the time of update.