Convert IDs and add basic interactions.
Turns out that | is used in some types of css selectors. See:
http://www.w3.org/TR/CSS2/selector.html#attribute-selectors
This conflicts with using | as a separator in our new style ids. In order to
deal with that I modified the getID and getNT methods for nucleotides and
interactions. These now convert the | to _. In addition I activated the basic
interaction controls.