Setup for toggling to airport view.
This adds a method in the model to extract the airport data from the DB if it
exists. It assumes that the data is stored in the database as a json encoded
string. It also modifies the controller to include this information in the data
and the 2D view to display toggles between the views if the airport diagram
exists.