Add a newer bootstrap button library
The newer version has a radio group of buttons which I need for displaying the
views, or at least I need that functionality and I don't want to reimplement
it. I've included it as a separate file without replacing the old one since I
don't know what will break. However, this is messy and should be changed. I'll
keep this file since it makes things work until Anton and I decide what to do,
weather to update, reorganize, or reimplement functionality.