Update controls to use newest library.
This fixes some bugs, most notably it actually displays 2D structures now.
However, there are still some problems.

1. The buttons that should toggle do not, for example consider the IL, HL, and
Click/Select buttons. This appears to be due to the fact that we are using an
outdated version of bootstrap-buttons.js. However, I am not going to upgrade
this yet. Without the newest version they do not show the expected toggle
behavior.

2. We cannot display 3D structures. This has something to do with a complaint
inside jmol.js, which I have no control over. The cause may be else where but I
haven't isolated it yet.

3. The circle is too small.

4. It does not start in airport view for those structures that have an airport
view.

Overall this commit is an improvment so I will add it.