Add a well for related structures.
This well shows all related structures and provides links to their 2D diagrams.
This is a useful thing Anton has requested and this is a basic version of it.
There are some problems with it. Notably the size of the jmol window above and
it do not match up. In addition I think the 2D diagram is now stretched so I'll
have to fix that.