Add basic view of motif positions to motif pages
We now generate a table of location -> loops for all known locations of loops
in motifs. This is put in a new tab 'Observed Locations' on the motif page. We
don't show the loops with an unknown position and I'm not sure if we should.
This is a test to browse some pages and see if I like the annotations.