Display number as well as sequence on hover
When we hover over a letter in the circular diagram we should also show the
number to allow for easy finding of particular nt. This makes such a
modification. It is a basic version and does have problems when hovering over
an interaction with two nearby nts the labels will overlap. In addition, this
modification should be added to the airport diagram as well, but is not yet.