Check for url parameter before setting view
We would like to link directly to all circular diagrams. The current resonable
default when seeing 2d's is to show an airport if possible. In order to provide
such links we add a check to examine the url parameter for a view=ciruclar
entry. If this happens we never set the view to airport.