Add a button for toggline long range interactions
This can turn on or off long range interactions. However this interacts with
the base pair family buttons. So it will only turn them off if the family
button is off as well as turning off the lr button. Not sure this is best
behavior but it is good enough for dev.