Add some js scripts
These scripts were loaded from rna.bgsu.edu without looking at the baseurl. This
was probably done so they can be easily shared across different sites, but this
means this site is not self-contained. This means the content is blocked when
running in the new test server. For this reason I'm adding the scripts to this
site.