First usage of modified nr_release_diff After multiple attempts, the best way to corral this complicated database query is within a MySQL stored procedure.  Given the same name as the table it replaces to avoid confusion.
The data was too complex to encapsulate as views.  The query was not amenable to CodeIgniter Active Record methods (CI attempted to add quotes in places that break the query).  The stored procedure approach avoids both of these pitfalls -- the database call via CI is much simpler, and the procedure is written to limit the amount of data under consideration, so that the query returns in a reasonable amount of time.