Attempt to fix popover hiding under jmol
Jmol sets the z-index to something pretty high so it tends to be above things
on the page. I'm not sure how to fix that, and if it will cause problems
elsewhere. Instead we can just set the popover z-index to be high enough that
it is not an issue.