Working chainbreak; restored conformation display Rewrote the query that calculates the chainbreak.  With the new database structure, it can be done in one database rount-trip step instead of two.
Along the line, I broke the conformation display (so that the conformation columns in the sequence variants view were not showing).  This issue is now fixed.

The page still exhibits a number of PHP errors.  At present, they are known to cluster in three areas of the code.  One source of the problem is non-equivalent unit_id values between ml_loop_positions and unit_info.  Further characterization and remediation steps are in progress but not reflected in this commit.