Modified query for PDB 2D Structure panel Since Blake deactivated the pipeline component that produced the pdb_best_chains_and_models table, we need to acquire that data (or its equivalent) from another source.
This query is a reasonable take on the question.  It's a bit more involved (because of the use of group_concat, and the multiple table join needed), but does produce data that resembles the original best_chains.

It may not be 100% equivalent.  Potential sources of error include the changes to the query, changes to the database, and missing data (known to throw a "Trying to get property of non-object" error when attempting to parse a NULL database response).

(cherry picked from commit 0b5a55cd63669872cd61a09ac93444db26d91a73)