Updated "most recent heatmap" logic Fixes a logic fault in the most recent heatmap SQL query (previous was not guaranteed to return the most recent release index).
Removes the logic to embed the nr_release_id value in the URL.  In
group meeting (2017-07-20), we assessed the issue of propagating
multiple versions of the heatmap data in the website cache.  Since
this is generally undesirable, we revert back to a single page per
set that will display the heatmap associated with the most recent
release.

Still includes some debugging statements for testing purposes.