Add related logic for alt_id and sym_op With the revised version of ins_code detection working, time to add the two related components.  Sym_op has already been observed to be affecting the data display, although alt_id hasn't (yet).
Exact form of representation is of course subject to change.  For
the first pass, I prepended each of the three "extra" unit_id
components with a different character.  Caret is bound to
ins_code, because that's how it is represented in the JSmol
display accompanying this presentation.  Underscore (alt_id) and
pipe (sym_op) were two other convenient characters (but not the
only options).